chore: port existing button styles to legacy files
