+++ src/core/styles.components.css
@@ -1,4 +1,5 @@
 @import "./styles/buttons.css";
+@import "./styles/legacy-buttons.css";
 @import "./styles/dropdowns.css";
 @import "./styles/forms.css";
 @import "./styles/layout.css";
